refactor: queries
