--- libs/service-portal/vehicles/src/queries/getUsersVehicleSearchLimit.ts
@@ -1,7 +1,0 @@
-import { gql } from '@apollo/client'
-
-export const GET_USERS_VEHICLES_SEARCH_LIMIT = gql`
-  query GetUsersVehiclesSearchLimit {
-    vehiclesSearchLimit
-  }
-`
